+++ lib/assets/javascripts/cartodb3/editor/editor-view.js
@@ -4,7 +4,7 @@ var Backbone = require('backbone');
 var AddWidgetsView = require('../components/modals/add-widgets/add-widgets-view');
 var createTextLabelsTabPane = require('../components/tab-pane/create-text-labels-tab-pane');
 var StackLayoutView = require('../components/stack-layout/stack-layout-view');
-var Header = require('./header.js');
+var Header = require('./editor-header.js');
 var TabPaneTemplate = require('./tab-pane.tpl');
 var EditorWidgetsView = require('./widgets/widgets-view');
 
@@ -57,8 +57,7 @@ module.exports = cdb.core.View.extend({
     }];
 
     var header = new Header({
-      title: this._visDefinitionModel.get('name'),
-      canGoBack: false
+      title: this._visDefinitionModel.get('name')
     });
 
     var tabPaneOptions = {
