+++ lib/assets/javascripts/cartodb3/editor/widgets/widgets-form/widget-header.js
@@ -1,19 +1,17 @@
 var cdb = require('cartodb.js');
-var template = require('./header.tpl');
+var template = require('./widget-header.tpl');
 
 module.exports = cdb.core.View.extend({
   initialize: function (opts) {
     if (!opts.title) throw new Error('title is required');
 
     this._title = opts.title;
-    this._canGoBack = opts.canGoBack;
   },
 
   render: function () {
     this.$el.html(
       template({
-        title: this._title,
-        canGoBack: this._canGoBack
+        title: this._title
       })
     );
     return this;
