+++ lib/assets/javascripts/cartodb3/editor/widgets/widgets-form/widget-header.tpl
@@ -1,0 +1,11 @@
+<div class="CDB-HeaderInfo">
+  <button class="u-rSpace--xl CDB-HeaderInfo-Back js-back">
+    <i class="CDB-IconFont CDB-IconFont-arrowPrev Size-large"></i>
+  </button>
+  <div class="CDB-HeaderInfo-Inner">
+    <div class="CDB-HeaderInfo-Title u-bSpace">
+      <h2 class="CDB-HeaderInfo-TitleText CDB-Text CDB-Size-huge is-light"><%- title %></h2>
+    </div>
+  </div>
+</div>
+
