+++ lib/assets/javascripts/cartodb3/editor/widgets/widgets-form/widgets-form-content-view.js
@@ -3,7 +3,7 @@ var WidgetsFormContentDataView = require('./widgets-form-content-data-view');
 var WidgetsFormContentStyleView = require('./widgets-form-content-style-view');
 var createTextLabelsTabPane = require('../../../components/tab-pane/create-text-labels-tab-pane');
 var TabPaneTemplate = require('./tab-pane.tpl');
-var Header = require('../../header.js');
+var Header = require('./widget-header.js');
 
 /**
  * View to render all necessary for the widget form
@@ -73,8 +73,7 @@ module.exports = cdb.core.View.extend({
     };
 
     var header = new Header({
-      title: this._widgetDefinitionModel.get('title'),
-      canGoBack: true
+      title: this._widgetDefinitionModel.get('title')
     });
 
     this.tabPaneView = createTextLabelsTabPane(this.tabPaneItems, options);
